+++ hobj/images/__init__.py
@@ -1,4 +1,4 @@
-from hobj.images.imagesets.highvar import MutatorHighVarImageset
-from hobj.images.imagesets.oneshot import MutatorOneShotImageset
-from hobj.images.imagesets.warmup import MutatorWarmupImageset
-from hobj.images.imagesets.probe import ProbeImageset
+from hobj.images.highvar import MutatorHighVarImageset
+from hobj.images.oneshot import MutatorOneShotImageset
+from hobj.images.warmup import MutatorWarmupImageset
+from hobj.images.probe import ProbeImageset
