+++ hobj/images/oneshot.py
